+++ CHANGELOG.md
@@ -5,3 +5,7 @@
 #### Gems
 
 - `Gem::Author` for authors info to be used primarily in gem specs.
+
+#### RSpec
+
+- API for method specs (`rspec/method`).